Module sui_adapter_v2::programmable_transactions::context

Structs

ExecutionContext

Maintains all runtime state specific to programmable transactions

Functions

load_type

Load type_tag to get a Type in the provided session. session’s linkage context may be reset after this operation, because during the operation, it may change when loading a struct.

load_type_from_struct